moved the stories component
